chore(db): move applied scripts
