+++ packages/db/prisma/seed.ts
@@ -1,5 +1,7 @@
 import { PrismaClient } from "@prisma/client";
 
+import permissions5 from "./aplied_scripts/new_seed_permission_01_09_2023";
+import permissions6 from "./aplied_scripts/new_seed_permission_02_09_2023";
 import agreementsScript from "./aplied_scripts/seed_agreements";
 import permissions4 from "./aplied_scripts/seed_edit_zone_permission_28_02_2023";
 import memberRole from "./aplied_scripts/seed_member_role";
@@ -8,7 +10,6 @@ import permissions2 from "./aplied_scripts/seed_new_permissions_21-01-2023";
 import permissions3 from "./aplied_scripts/seed_new_permissions_22-01-2023";
 import permissions0 from "./aplied_scripts/seed_new_permissions_27_12_2022";
 import seedRoles from "./aplied_scripts/seed_roles";
-import permissions5 from "./new_seed_permission_01_09_2023";
 
 const prisma = new PrismaClient();
 
@@ -34,6 +35,7 @@ async function main() {
   await permissions3();
   await permissions4();
   await permissions5();
+  await permissions6();
 }
 
 main()
